Move DetectLockSupport and lock detection logic into LockDirectory
- DetectLockSupport and its four Test* helpers move from CachedDriveInfo
  to LockDirectory, where they belong alongside GetLockDirectory.
- GetLockDirectory now uses drive?.LockSupport ?? DetectLockSupport(...)
  so drive heuristics act as a shortcut; detection runs whenever the
  drive has no opinion (null LockSupport) or there is no matching drive.
- Tests updated: DriveWithCapability uses "unknown-fs" when capability
  is null so LockSupport correctly returns null and triggers detection;
  the empty-drives test is updated to reflect that detection now runs
  (and succeeds) rather than returning Unsupported unconditionally.
